Call #to_s on :protocol and :state before saving
This results in Strings being passed around so the framework's
 #set_property can properly remove duplicates.

Originally we were sending Symbols, which was fine as they got
serialised as Strings in the database. However on their way up (e.g. if
you upload the same file twice), the framework is trying to compare the
values from the database (strings) with the ones the plugin is sending
(symbols) and since they don't match, you end with duplicates (one entry
for strings, one for symbols) that eventually turn to complete
duplicates once everything is serialised as strings in the database.

See:
  https://github.com/dradis/dradis-ce/issues/43